chore(core): add color module
Move color functionality to a dedicated module.
